# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gupta Enterprises E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (user's own, or all for admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status (Admin)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/assign-agent:
    post:
      operationId: assignDeliveryAgent
      tags: [orders]
      summary: Assign delivery agent to order (Admin)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignAgentBody"
      responses:
        "200":
          description: Order with assigned agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List all coupons (Admin)
      responses:
        "200":
          description: List of coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create a coupon (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponBody"
      responses:
        "201":
          description: Created coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{couponId}:
    put:
      operationId: updateCoupon
      tags: [coupons]
      summary: Update a coupon (Admin)
      parameters:
        - name: couponId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCouponBody"
      responses:
        "200":
          description: Updated coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete a coupon (Admin)
      parameters:
        - name: couponId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCouponBody"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidationResult"

  # REVIEWS
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        firebaseUid:
          type: string
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [USER, ADMIN, DELIVERY_AGENT]
        photoUrl:
          type: string
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/Address"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - createdAt
    Address:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        lat:
          type: number
        lng:
          type: number
        isDefault:
          type: boolean
      required:
        - id
        - street
        - city
        - state
        - pincode
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        actualPrice:
          type: number
        sellingPrice:
          type: number
        discount:
          type: number
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - actualPrice
        - sellingPrice
        - stock
        - images
        - isActive
        - createdAt
...
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        productCount:
          type: integer
      required:
        - id
        - name
        - slug
...
    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        user:
          $ref: "#/components/schemas/User"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        status:
          type: string
          enum: [PENDING, CONFIRMED, PROCESSING, PACKED, OUT_FOR_DELIVERY, DELIVERED, CANCELLED, PICKUP_READY, PICKED_UP]
        deliveryType:
          type: string
          enum: [DELIVERY, PICKUP]
        deliveryAddress:
          $ref: "#/components/schemas/Address"
        subtotal:
          type: number
        discount:
          type: number
        couponDiscount:
          type: number
        total:
          type: number
        couponCode:
          type: string
        deliveryAgentId:
          type: string
        deliveryAgent:
          $ref: "#/components/schemas/User"
        estimatedDelivery:
          type: string
          format: date-time
        paymentStatus:
          type: string
          enum: [PENDING, PAID, FAILED, REFUNDED]
        paymentMethod:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - items
        - status
        - deliveryType
        - subtotal
        - total
        - paymentStatus
        - createdAt
        - updatedAt
    OrderItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        price:
          type: number
        total:
          type: number
      required:
        - id
        - productId
        - quantity
        - price
        - total
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - orders
        - total
        - page
        - limit
        - totalPages
    Coupon:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        description:
          type: string
        discountType:
          type: string
          enum: [FLAT, PERCENTAGE]
        discountValue:
          type: number
        minOrderValue:
          type: number
        maxDiscount:
          type: number
        isFirstOrder:
          type: boolean
        isActive:
          type: boolean
        usageLimit:
          type: integer
        usageCount:
          type: integer
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - code
        - discountType
        - discountValue
        - isActive
        - createdAt
    CouponValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        discount:
          type: number
        message:
          type: string
        coupon:
          $ref: "#/components/schemas/Coupon"
      required:
        - valid
        - message
...
    CreateOrderBody:
      type: object
      properties:
        deliveryType:
          type: string
          enum: [DELIVERY, PICKUP]
        deliveryAddress:
          $ref: "#/components/schemas/Address"
        paymentMethod:
          type: string
        notes:
          type: string
        pickupDate:
          type: string
          format: date-time
      required:
        - deliveryType
        - paymentMethod
    UpdateOrderBody:
      type: object
      properties:
        status:
          type: string
          enum: [PENDING, CONFIRMED, PROCESSING, PACKED, OUT_FOR_DELIVERY, DELIVERED, CANCELLED, PICKUP_READY, PICKED_UP]
        estimatedDelivery:
          type: string
          format: date-time
      required:
        - status
    AssignAgentBody:
      type: object
      properties:
        agentId:
          type: string
      required:
        - agentId
    CreateCouponBody:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
        discountType:
          type: string
          enum: [FLAT, PERCENTAGE]
        discountValue:
          type: number
        minOrderValue:
          type: number
        maxDiscount:
          type: number
        isFirstOrder:
          type: boolean
        usageLimit:
          type: integer
        expiresAt:
          type: string
          format: date-time
      required:
        - code
        - discountType
        - discountValue
    UpdateCouponBody:
      type: object
      properties:
        description:
          type: string
        discountValue:
          type: number
        minOrderValue:
          type: number
        maxDiscount:
          type: number
        isFirstOrder:
          type: boolean
        isActive:
          type: boolean
        usageLimit:
          type: integer
        expiresAt:
          type: string
          format: date-time
    ValidateCouponBody:
      type: object
      properties:
        code:
          type: string
        orderTotal:
          type: number
      required:
        - code
        - orderTotal
...