# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gupta Enterprises E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Address } from "./address";
import type { CreateOrderBodyDeliveryType } from "./createOrderBodyDeliveryType";

export interface CreateOrderBody {
  deliveryType: CreateOrderBodyDeliveryType;
  deliveryAddress?: Address;
  paymentMethod: string;
  notes?: string;
  pickupDate?: Date;
}