# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gupta Enterprises E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Address } from "./address";
import type { OrderDeliveryType } from "./orderDeliveryType";
import type { OrderItem } from "./orderItem";
import type { OrderPaymentStatus } from "./orderPaymentStatus";
import type { OrderStatus } from "./orderStatus";
import type { User } from "./user";

export interface Order {
  id: string;
  userId: string;
  user?: User;
  items: OrderItem[];
  status: OrderStatus;
  deliveryType: OrderDeliveryType;
  deliveryAddress?: Address;
  subtotal: number;
  discount?: number;
  couponDiscount?: number;
  total: number;
  couponCode?: string;
  deliveryAgentId?: string;
  deliveryAgent?: User;
  estimatedDelivery?: Date;
  paymentStatus: OrderPaymentStatus;
  paymentMethod?: string;
  notes?: string;
  createdAt: Date;
  updatedAt: Date;
}